+++ backend-api/src/entities/report_social.entity.ts
@@ -5,7 +5,7 @@ import { Report } from './reports.entity';
 
 @Entity('report_social')
 export class ReportSocial {
-  @PrimaryColumn({ type: 'bigint' })
+  @PrimaryColumn({ type: 'int' })
   report_id: number;
 
   @OneToOne(() => Report, { onDelete: 'CASCADE' })
